+++ WordPress/src/main/java/org/wordpress/android/ui/posts/services/PostUploadEvents.java
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.posts;
+package org.wordpress.android.ui.posts.services;
 
 public class PostUploadEvents {
     public static class PostUploadSucceed {
